- Create FlaskThread to handle thread inside Flask app context - Split setting.py, Flask app.py, and gunicorn server.py
