--- src/envs.py
@@ -1,1 +1,0 @@
-DATA_DIR_ENV = 'RUBIX_LORA_DATA'
